+++ @doc/roadmap.md
@@ -14,16 +14,15 @@
 - [x] 批量修改(统一设置模式+单个设置模式)、批量删除
 - [x] request的version设置
 - [x] 增加 functions , 用于对数据的处理 (字典转换、id转name等操作)
-- [ ] ...
+
 
 ## [ ] 摸石头 阶段
 - [ ] 开发、测试环境下记录get请求记录
-- [ ] 可自定义get中查询节点的实现 (即从缓存或者其他数据源查询数据)
-- [ ] 可自定义非开放请求的具体实现
+- [x] 可自定义get中查询节点的实现 (即从缓存或者其他数据源查询数据)
+- [x] 可自定义非开放请求的具体实现
+- [x] 提供 web UI 用于 _access 和 _request 的管理
 
 ## [ ] xxx 阶段
-
-- [x] 提供 web UI 用于 _access 和 _request 的管理
 - [ ] get请求的复杂度计算, 超过复杂度则拒绝执行(树节点计算)
 - [ ] 完善 apijson 规范的兼容实现
 
@@ -47,8 +46,9 @@
     - [x] get只有access中定义的才能访问
     - [x] 非get操作则必须与request指定一致才可请求
     - [x] 基于角色控制
+
 - [x] 远程函数
-- [ ] 错误提示
-- [ ] 查询节点 自定义查询数据
+- [x] 查询节点 自定义查询数据
 - [x] 字段限制
+- [ ] 错误提示完善
 - [ ] 请求结构复杂度限制
