+++ config/rowkeygen.go
@@ -1,15 +1,16 @@
-package action
+package config
 
 import (
 	"context"
 	"github.com/gogf/gf/v2/frame/g"
+	"github.com/samber/lo"
 )
 
 type RowKeyGenFuncHandler func(ctx context.Context, genParam g.Map, table string, data g.Map) (g.Map, error)
 
 var rowKeyGenFuncMap = map[string]RowKeyGenFuncHandler{}
 
-func rowKeyGen(ctx context.Context, genFuncName string, table string, data g.Map) (g.Map, error) {
+func RowKeyGen(ctx context.Context, genFuncName string, table string, data g.Map) (g.Map, error) {
 	if f, exists := rowKeyGenFuncMap[genFuncName]; exists {
 		return f(ctx, g.Map{}, table, data)
 	}
@@ -19,4 +20,8 @@ func rowKeyGen(ctx context.Context, genFuncName string, table string, data g.Map
 
 func RowKeyGenFunc(name string, f RowKeyGenFuncHandler) {
 	rowKeyGenFuncMap[name] = f
+}
+
+func RowKeyGenList() []string {
+	return lo.Keys(rowKeyGenFuncMap)
 }
