+++ framework/gf_orm/query.go
@@ -3,11 +3,10 @@ package gf_orm
 import (
 	"context"
 	"github.com/glennliao/apijson-go/config"
+	"github.com/glennliao/apijson-go/config/db"
+	"github.com/glennliao/apijson-go/config/executor"
 	"github.com/glennliao/apijson-go/consts"
-	"github.com/glennliao/apijson-go/db"
-	"github.com/glennliao/apijson-go/query/executor"
 	"github.com/gogf/gf/v2/database/gdb"
-	"github.com/gogf/gf/v2/errors/gerror"
 	"github.com/gogf/gf/v2/frame/g"
 	"github.com/gogf/gf/v2/util/gconv"
 	"github.com/samber/lo"
@@ -349,85 +348,7 @@ func (e *SqlExecutor) One() (g.Map, error) {
 	return one.Map(), err
 }
 
-func Insert(ctx context.Context, table string, data any) (int64, int64, error) {
-
-	ret, err := g.DB().Insert(ctx, table, data)
-	if err != nil {
-		return 0, 0, err
-	}
-	id, err := ret.LastInsertId()
-	if err != nil {
-		return 0, 0, err
-	}
-
-	count, err := ret.RowsAffected()
-	return id, count, nil
-}
-
-func Update(ctx context.Context, table string, data g.Map, where g.Map) (int64, error) {
-
-	m := g.DB().Model(table).Ctx(ctx)
-
-	for k, v := range where {
-		if strings.HasSuffix(k, "{}") {
-			if vStr, ok := v.(string); ok {
-				if vStr == "" {
-					return 0, gerror.New("where的值不能为空")
-				}
-			}
-			m = m.WhereIn(k[0:len(k)-2], v)
-			delete(where, k)
-			continue
-		}
-		if v.(string) == "" || v == nil { //暂只处理字符串为空的情况
-			return 0, gerror.New("where的值不能为空")
-		}
-	}
-
-	_ret, err := m.Where(where).Update(data)
-	if err != nil {
-		return 0, err
-	}
-
-	count, err := _ret.RowsAffected()
-	if err != nil {
-		return 0, err
-	}
-
-	return count, err
-
-}
-
-func Delete(ctx context.Context, table string, where g.Map) (int64, error) {
-
-	if len(where) == 0 {
-		return 0, gerror.New("where不能为空")
-	}
-
-	m := g.DB().Model(table).Ctx(ctx)
-
-	for k, v := range where {
-		if strings.HasSuffix(k, "{}") {
-			m = m.WhereIn(k[0:len(k)-2], v)
-			delete(where, k)
-			continue
-		}
-		if v.(string) == "" || v == nil { //暂只处理字符串为空的情况
-			return 0, gerror.New("where的值不能为空")
-		}
-	}
-
-	_ret, err := m.Where(where).Delete()
-	if err != nil {
-		return 0, err
-	}
-
-	count, err := _ret.RowsAffected()
-
-	if err != nil {
-		return 0, err
-	}
-
-	return count, err
-
+// init 暂先自动注册,后续改成可手动配置
+func init() {
+	executor.RegQueryExecutor("default", New)
 }
