+++ go.mod
@@ -4,6 +4,7 @@ go 1.18
 
 require (
 	github.com/gogf/gf/v2 v2.2.5
+	github.com/iancoleman/orderedmap v0.2.0
 	github.com/samber/lo v1.33.0
 )
